+++ src/Plugin/WebformHandler/WebformHandlerSF1601.php
@@ -246,6 +246,7 @@ final class WebformHandlerSF1601 extends WebformHandlerBase {
       'os2forms_nemid_company_cvr',
       'os2forms_nemid_company_cvr_fetch_data',
       'os2forms_nemid_cpr',
+      'os2forms_person_lookup',
       // @todo Remove these when we remove the elements.
       'cpr_element',
       'cpr_value_element',